backend/requirements.txt
Authlib==1.2.0
certifi==2021.10.8
cffi==1.15.1
charset-normalizer==2.0.12
click==8.0.4
colorama==0.4.4
cryptography==40.0.2
Flask==2.0.3
Flask-Cors==3.0.10
flask-marshmallow==0.14.0
Flask-SQLAlchemy==2.5.1
greenlet==1.1.2
idna==3.3
itsdangerous==2.1.0
Jinja2==3.0.3
MarkupSafe==2.1.0
marshmallow==3.14.1
marshmallow-sqlalchemy==0.27.0
mysqlclient==2.1.0
pycparser==2.21
PyJWT==2.7.0
python-dotenv==1.0.0
six==1.16.0
SQLAlchemy==1.4.32
urllib3==1.26.8
Werkzeug==2.0.3
